refactor: align list page exports
