+++ resources/js/src/pages/dashboard/users/list/index.tsx
@@ -63,7 +63,7 @@ const metadata = { title: `Users | Dashboard - ${CONFIG.appName}` };
 
 // ----------------------------------------------------------------------
 
-export default function Index({ users, roles }: Props) {
+export default function List({ users, roles }: Props) {
   const { __ } = useLang();
   const { props } = usePage<PageProps>();
   const csrfToken = props.csrf_token;
