+++ resources/js/src/pages/dashboard/users/view/index.tsx
@@ -1,0 +1,11 @@
+import { CONFIG } from 'src/global-config';
+
+export default function View() {
+  return (
+    <>
+      <title>{`View User | Dashboard - ${CONFIG.appName}`}</title>
+      <div>TODO: Implement user view page</div>
+    </>
+  );
+}
+
